# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FocusFlow API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ProjectSession {
  id: string;
  title: string;
  description: string;
  recommended_agent: string;
  tech_stack_rules: string;
  win_condition: string;
  deliverables: string[];
  dependencies: string[];
}

export interface ProjectState {
  goal: string;
  outline: string[];
  sessions: ProjectSession[];
  obsidian_markdown: string;
}

export interface Project {
  id: number;
  title: string;
  projectState: ProjectState;
  completedSessions: string[];
  createdAt: string;
  updatedAt: string;
}

export interface ProjectMessage {
  id: number;
  projectId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ProjectWithMessages {
  id: number;
  title: string;
  projectState: ProjectState;
  completedSessions: string[];
  createdAt: string;
  updatedAt: string;
  messages: ProjectMessage[];
}

export interface ProjectInput {
  title: string;
}

export interface ProjectUpdate {
  title?: string;
  projectState?: ProjectState;
  completedSessions?: string[];
}

export interface ProjectMessageInput {
  content: string;
}

export interface ApiError {
  error: string;
}