# lib/api-zod/src/generated/types/projectMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FocusFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectMessage {
  id: number;
  projectId: number;
  role: string;
  content: string;
  createdAt: Date;
}