# lib/api-zod/src/generated/types/projectSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FocusFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectSession {
  id: string;
  title: string;
  description: string;
  recommended_agent: string;
  tech_stack_rules: string;
  win_condition: string;
  deliverables: string[];
  dependencies: string[];
}