# lib/api-zod/src/generated/types/projectState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FocusFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectSession } from "./projectSession";

export interface ProjectState {
  goal: string;
  outline: string[];
  sessions: ProjectSession[];
  obsidian_markdown: string;
}