# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FocusFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectState } from "./projectState";

export interface ProjectUpdate {
  title?: string;
  projectState?: ProjectState;
  completedSessions?: string[];
}