# lib/api-zod/src/generated/types/projectWithMessages.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FocusFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectMessage } from "./projectMessage";
import type { ProjectState } from "./projectState";

export interface ProjectWithMessages {
  id: number;
  title: string;
  projectState: ProjectState;
  completedSessions: string[];
  createdAt: Date;
  updatedAt: Date;
  messages: ProjectMessage[];
}